% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissever-package.R
\name{dissever-package}
\alias{dissever-package}
\title{dissever}
\description{
Dissever: algorithm for spatial downscaling
}
\details{
Algorithm described in: [DOI: 10.1016/j.cageo.2011.08.021]
}
\author{
Brendan Malone (brendan.malone@sydney.edu.au);
}

