% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissever.R
\name{print.dissever}
\alias{print.dissever}
\title{Prints the performance of the dissever procedure}
\usage{
\method{print}{dissever}(x, ...)
}
\arguments{
\item{x}{object of class \code{dissever}, output from the \code{dissever} function}

\item{...}{Additional arguments passed to print}
}
\description{
Prints the performance of the model used to do the dissever procedure.
}
\author{
Pierre Roudier
}

