% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissever.R
\name{summary.dissever}
\alias{summary.dissever}
\title{Prints summary of the model used in the dissever procedure}
\usage{
\method{summary}{dissever}(object, ...)
}
\arguments{
\item{object}{object of class \code{dissever}, output from the \code{dissever} function
#'}

\item{...}{Additional arguments passed to summary}
}
\description{
Prints summary of the model used in the dissever procedure.
}
\author{
Pierre Roudier
}

